refactored / restructured the app
