+++ Slack/responses/checkinResponses.js
