+++ Slack/responses/profileResponses.js
@@ -2,7 +2,7 @@
  * Created by Vampiire on 8/9/17.
  */
 
-const userProfile = require('../database/profileModel').userProfile;
+const userProfile = require('../../database/profileModel').userProfile;
 
 
 // ------------------- CORE EXPORTS ------------------- //
