+++ Slack/routes/slashInteractive.js
@@ -2,7 +2,7 @@
 const express = require('express');
 const router = module.exports = express.Router();
 
-const tools = require('../../tools/exporter');
+const tools = require('../tools/exporter');
 
 // handles interactive message responses / submissions
 router.post('/', (req, res) => {
