+++ Slack/routes/slashUpdate.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = module.exports = express.Router();
 
-const tools = require('../../tools/exporter');
-const updateResponse = require('../../responses/updateResponses');
+const tools = require('../tools/exporter');
+const updateResponse = require('../responses/updateResponses');
 
 const userProfile = require('../../database/profileModel').userProfile;
