+++ Slack/tools/valStringer.js
