+++ Slack/tools/verify.js
